Fix for OSError: [Errno 121] Remote I/O error
Adding a delay of 1 sec after establishing the i2c bus prevents this error.  Seems you need to allow the i2c bus to establish before sending a command, otherwise you get this error.  Tried reducing the delay time but 1 sec seems to be the right length.  Tested on an RPi3.