Fix/22 io error long cmds (#23)
* feat: Add function for parsing the firmware version info into a named tuple.

* fix: Fix io errors due to driver not polling ACK frame correctly. Fixes #22

Long commands do not ack immediately so we are supposed to poll for the ack frame. quick2wire throws an error if the read is not acked though so we need to catch those errors and poll again after a brief wait. If we timeout we return an error code instead of raising an exception.

* doc: Update README to indicate using the 5v pin on the raspberry pi is safe.
Bump version minor.
Correct email in packaging step.